restructured utils to get used function for loop passage function with context outside the class
